Replace size_t with int in a few media classes.
This addressed TODOs in DataSource::Read() and AudioRendererAlgorithmBase, which led to converting Buffer and SeekableBuffer as well.

Review URL: https://chromiumcodereview.appspot.com/9854031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129140 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/base/data_buffer.cc
@@ -12,14 +12,14 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace media {
 
-DataBuffer::DataBuffer(scoped_array<uint8> buffer, size_t buffer_size)
+DataBuffer::DataBuffer(scoped_array<uint8> buffer, int buffer_size)
     : Buffer(base::TimeDelta(), base::TimeDelta()),
       data_(buffer.Pass()),
       buffer_size_(buffer_size),
       data_size_(buffer_size) {
 }
 
-DataBuffer::DataBuffer(size_t buffer_size)
+DataBuffer::DataBuffer(int buffer_size)
     : Buffer(base::TimeDelta(), base::TimeDelta()),
       data_(new uint8[buffer_size]),
       buffer_size_(buffer_size),
@@ -34,8 +34,8 @@ DataBuffer::DataBuffer(size_t buffer_size)
 DataBuffer::~DataBuffer() {}
 
 scoped_refptr<DataBuffer> DataBuffer::CopyFrom(const uint8* data,
-                                               size_t data_size) {
-  size_t padding_size = 0;
+                                               int data_size) {
+  int padding_size = 0;
 #if !defined(OS_ANDROID)
   // Why FF_INPUT_BUFFER_PADDING_SIZE? FFmpeg assumes all input buffers are
   // padded with this value.
@@ -54,7 +54,7 @@ const uint8* DataBuffer::GetData() const {
   return data_.get();
 }
 
-size_t DataBuffer::GetDataSize() const {
+int DataBuffer::GetDataSize() const {
   return data_size_;
 }
 
@@ -67,12 +67,12 @@ uint8* DataBuffer::GetWritableData() {
 }
 
 
-void DataBuffer::SetDataSize(size_t data_size) {
+void DataBuffer::SetDataSize(int data_size) {
   DCHECK_LE(data_size, buffer_size_);
   data_size_ = data_size;
 }
 
-size_t DataBuffer::GetBufferSize() const {
+int DataBuffer::GetBufferSize() const {
   return buffer_size_;
 }
